requirements.txt
streamlit
langchain==0.3.26
langchain-community==0.3.26
langchain-pinecone==0.2.8
langchain-openai==0.3.24
langchain-text-splitters
pypdf==5.6.1
sentence-transformers==4.1.0
python-dotenv==1.1.0
pinecone-client
